requirements.txt
matplotlib==3.8.3
opencv-python==4.9.0.80
Pillow==9.4.0
scikit-image==0.19.3
scikit-learn==1.4.1.post1
scipy==1.13.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.66.2
